refactor: Better separation of generic collection view code and implementation specifics
This change moves a lot of the reusable selection management code into an `NSCollectionView` extension and leaves only the implementation-specific code in `InteractiveCollectionView`. It also extracts a couple of utilities and renames a couple of APIs to match the implementation in SelectableCollectionView which we're currently tracking.
