+++ Folders/Views/InteractiveCollectionView.swift
@@ -23,53 +23,19 @@
 import AppKit
 import Carbon
 
-protocol InteractiveCollectionViewDelegate: NSObject {
+protocol CollectionViewInteractionDelegate: NSObject {
 
-    func customCollectionView(_ customCollectionView: InteractiveCollectionView,
-                              contextMenuForSelection selection: IndexSet) -> NSMenu?
-    func customCollectionView(_ customCollectionView: InteractiveCollectionView,
-                              didDoubleClickSelection selection: Set<IndexPath>)
-    func customCollectionViewShowPreview(_ customCollectionView: InteractiveCollectionView)
+    func collectionView(_ collectionView: InteractiveCollectionView, contextMenuForSelection selection: Set<IndexPath>) -> NSMenu?
+    func collectionView(_ collectionView: InteractiveCollectionView, didDoubleClickSelection selection: Set<IndexPath>)
+    func collectionViewShowPreview(_ collectionView: InteractiveCollectionView)
 
 }
 
 class InteractiveCollectionView: NSCollectionView {
 
-    struct NavigationResult {
-        let nextIndexPath: IndexPath
-        let intermediateIndexPaths: [IndexPath]
-
-        init?(nextIndexPath: IndexPath?, intermediateIndexPaths: [IndexPath] = []) {
-            guard let nextIndexPath else {
-                return nil
-            }
-            self.nextIndexPath = nextIndexPath
-            self.intermediateIndexPaths = intermediateIndexPaths
-        }
-    }
-
-    weak var interactionDelegate: InteractiveCollectionViewDelegate?
+    weak var interactionDelegate: CollectionViewInteractionDelegate?
 
     var cursor: IndexPath?
-
-    override func menu(for event: NSEvent) -> NSMenu? {
-
-        // Update the selection if necessary.
-        let point = convert(event.locationInWindow, from: nil)
-        if let indexPath = indexPathForItem(at: point) {
-            if !selectionIndexPaths.contains(indexPath) {
-                selectionIndexPaths = [indexPath]
-            }
-        } else {
-            selectionIndexPaths = []
-        }
-
-        if let menu = interactionDelegate?.customCollectionView(self, contextMenuForSelection: selectionIndexes) {
-            return menu
-        }
-
-        return super.menu(for: event)
-    }
 
     func fixupSelection(direction: NavigationDirection) {
 
@@ -89,6 +55,25 @@ class InteractiveCollectionView: NSCollectionView {
             }
             return
         }
+    }
+
+    override func menu(for event: NSEvent) -> NSMenu? {
+
+        // Update the selection if necessary.
+        let point = convert(event.locationInWindow, from: nil)
+        if let indexPath = indexPathForItem(at: point) {
+            if !selectionIndexPaths.contains(indexPath) {
+                selectionIndexPaths = [indexPath]
+            }
+        } else {
+            selectionIndexPaths = []
+        }
+
+        if let menu = interactionDelegate?.collectionView(self, contextMenuForSelection: selectionIndexPaths) {
+            return menu
+        }
+
+        return super.menu(for: event)
     }
 
     override func mouseDown(with event: NSEvent) {
@@ -112,7 +97,7 @@ class InteractiveCollectionView: NSCollectionView {
 
                     if indexPath > cursor {
                         var indexPaths = Set<IndexPath>()
-                        for i in self.indexPaths(following: cursor, direction: .forwards) {
+                        for i in self.indexPathSequence(following: cursor, direction: .forwards) {
                             guard i <= indexPath else {
                                 break
                             }
@@ -126,7 +111,7 @@ class InteractiveCollectionView: NSCollectionView {
 
                     } else {
                         var indexPaths = Set<IndexPath>()
-                        for i in self.indexPaths(following: cursor, direction: .backwards) {
+                        for i in self.indexPathSequence(following: cursor, direction: .backwards) {
                             guard i >= indexPath else {
                                 break
                             }
@@ -189,164 +174,9 @@ class InteractiveCollectionView: NSCollectionView {
 
         // Handle double-click.
         if event.clickCount > 1, !selectionIndexPaths.isEmpty {
-            interactionDelegate?.customCollectionView(self, didDoubleClickSelection: selectionIndexPaths)
+            interactionDelegate?.collectionView(self, didDoubleClickSelection: selectionIndexPaths)
             return
         }
-    }
-
-    func indexPaths(following indexPath: IndexPath, direction: SequenceDirection) -> IndexPathSequence {
-        return IndexPathSequence(collectionView: self, indexPath: indexPath, direction: direction)
-    }
-
-    func firstIndexPath() -> IndexPath? {
-        return self.indexPath(after: IndexPath(item: -1, section: 0))
-    }
-
-    func lastIndexPath() -> IndexPath? {
-        return self.indexPath(before: IndexPath(item: 0, section: numberOfSections))
-    }
-
-    func indexPath(before indexPath: IndexPath) -> IndexPath? {
-
-        // Try decrementing the item...
-        if indexPath.item - 1 >= 0 {
-            return IndexPath(item: indexPath.item - 1, section: indexPath.section)
-        }
-
-        // Try decrementing the section...
-        var nextSection = indexPath.section
-        while true {
-            nextSection -= 1
-            guard nextSection >= 0 else {
-                return nil
-            }
-            let numberOfItems = numberOfItems(inSection: nextSection)
-            if numberOfItems > 0 {
-                return IndexPath(item: numberOfItems - 1, section: nextSection)
-            }
-        }
-
-    }
-
-    func indexPath(after indexPath: IndexPath) -> IndexPath? {
-
-        // Try incrementing the item...
-        if indexPath.item + 1 < numberOfItems(inSection: indexPath.section) {
-            return IndexPath(item: indexPath.item + 1, section: indexPath.section)
-        }
-
-        // Try incrementing the section...
-        var nextSection = indexPath.section
-        while true {
-            nextSection += 1
-            guard nextSection < numberOfSections else {
-                return nil
-            }
-            if numberOfItems(inSection: nextSection) > 0 {
-                return IndexPath(item: 0, section: nextSection)
-            }
-        }
-
-    }
-
-    func indexPath(following indexPath: IndexPath, direction: SequenceDirection, distance: Int = 1) -> IndexPath? {
-        var indexPath: IndexPath? = indexPath
-        for _ in 0..<distance {
-            guard let testIndexPath = indexPath else {
-                return nil
-            }
-            switch direction {
-            case .forwards:
-                indexPath = self.indexPath(after: testIndexPath)
-            case .backwards:
-                indexPath = self.indexPath(before: testIndexPath)
-            }
-        }
-        return indexPath
-    }
-
-    func closestIndexPath(toIndexPath indexPath: IndexPath, direction: NavigationDirection) -> NavigationResult? {
-
-        guard let layout = collectionViewLayout else {
-            return nil
-        }
-
-        let threshold = 20.0
-        let attributesForCurrentItem = layout.layoutAttributesForItem(at: indexPath)
-        let currentItemFrame = attributesForCurrentItem?.frame ?? .zero
-        let targetPoint: CGPoint
-        let indexPaths: IndexPathSequence
-        switch direction {
-        case .up:
-            targetPoint = CGPoint(x: currentItemFrame.midX, y: currentItemFrame.minY - threshold)
-            indexPaths = self.indexPaths(following: indexPath, direction: .backwards)
-        case .down:
-            targetPoint = CGPoint(x: currentItemFrame.midX, y: currentItemFrame.maxY + threshold)
-            indexPaths = self.indexPaths(following: indexPath, direction: .forwards)
-        case .left:
-            targetPoint = CGPoint(x: currentItemFrame.minX - threshold, y: currentItemFrame.midY)
-            indexPaths = self.indexPaths(following: indexPath, direction: .backwards)
-        case .right:
-            targetPoint = CGPoint(x: currentItemFrame.maxX + threshold, y: currentItemFrame.midY)
-            indexPaths = self.indexPaths(following: indexPath, direction: .forwards)
-        }
-
-        // This takes a really simple approach that either walks forwards or backwards through the cells to find the
-        // next cell. It will fail hard on sparsely packed layouts or layouts which place elements randomly but feels
-        // like a reasonable limitation given the current planned use-cases.
-        //
-        // A more flexible implementation might compute the vector from our current item to the test item and select one
-        // with the lowest magnitude closest to the requested direction. It might also be possible to use this approach
-        // to do wrapping more 'correctly'.
-        //
-        // Seeking should probably also be limited to a maximum nubmer of test items to avoid walking thousands of items
-        // if no obvious match is found.
-
-        var intermediateIndexPaths: [IndexPath] = []
-        for indexPath in indexPaths {
-            if let attributes = layout.layoutAttributesForItem(at: indexPath),
-               attributes.frame.contains(targetPoint) {
-                return NavigationResult(nextIndexPath: indexPath, intermediateIndexPaths: intermediateIndexPaths)
-            }
-            intermediateIndexPaths.append(indexPath)
-        }
-        return nil
-    }
-
-    func nextIndex(_ direction: NavigationDirection, indexPath: IndexPath?) -> NavigationResult? {
-
-        // This implementation makes some assumptions that will work with packed grid-like layouts but are unlikely to
-        // work well with sparsely packed layouts or irregular layouts. Specifically:
-        //
-        // - Left/Right directions are always assumed to selection the previous or next index paths by item and section.
-        //
-        // - Up/Down will seek through the index paths in order and return the index path of the first item which
-        //   contains a point immediately above or below the starting index path.
-
-        guard let indexPath else {
-            switch direction.sequenceDirection {
-            case .forwards:
-                return NavigationResult(nextIndexPath: firstIndexPath())
-            case .backwards:
-                return NavigationResult(nextIndexPath: lastIndexPath())
-            }
-        }
-
-        switch direction {
-        case .up, .down:
-            return closestIndexPath(toIndexPath: indexPath, direction: direction)
-        case .left:
-            return NavigationResult(nextIndexPath: self.indexPath(before: indexPath))
-        case .right:
-            return NavigationResult(nextIndexPath: self.indexPath(after: indexPath))
-        }
-    }
-
-    func isSelected(_ indexPath: IndexPath?) -> Bool {
-        guard let indexPath else {
-            return false
-        }
-        return selectionIndexPaths.contains(indexPath)
     }
 
     override func keyDown(with event: NSEvent) {
@@ -359,7 +189,7 @@ class InteractiveCollectionView: NSCollectionView {
             // TODO: Consider fixing up the selection here too. (Finder does a cunning thing where if you have a
             //       selection of size greater than one, the preview cycles the items in the selection, but if there's
             //       only one item, the arrow keys directly manipulate the items in the grid.)
-            interactionDelegate?.customCollectionViewShowPreview(self)
+            interactionDelegate?.collectionViewShowPreview(self)
             return
         }
 
@@ -367,7 +197,7 @@ class InteractiveCollectionView: NSCollectionView {
         if event.modifierFlags.contains(.command),
            event.keyCode == kVK_DownArrow {
             if !selectionIndexPaths.isEmpty {
-                interactionDelegate?.customCollectionView(self, didDoubleClickSelection: selectionIndexPaths)
+                interactionDelegate?.collectionView(self, didDoubleClickSelection: selectionIndexPaths)
             }
             return
         }
